+++ packages/core/src/js/tracing/timetodisplay.tsx
@@ -1,12 +1,13 @@
 import type { Span,StartSpanOptions  } from '@sentry/core';
 import { fill, getActiveSpan, getSpanDescendants, logger, SEMANTIC_ATTRIBUTE_SENTRY_ORIGIN, SPAN_STATUS_ERROR, SPAN_STATUS_OK, spanToJSON, startInactiveSpan } from '@sentry/core';
 import * as React from 'react';
+import { useState } from 'react';
 
 import { isTurboModuleEnabled } from '../utils/environment';
 import { SPAN_ORIGIN_AUTO_UI_TIME_TO_DISPLAY, SPAN_ORIGIN_MANUAL_UI_TIME_TO_DISPLAY } from './origin';
 import { getRNSentryOnDrawReporter, nativeComponentExists } from './timetodisplaynative';
-import type {RNSentryOnDrawNextFrameEvent } from './timetodisplaynative.types';
-import { setSpanDurationAsMeasurement } from './utils';
+import type { RNSentryOnDrawNextFrameEvent } from './timetodisplaynative.types';
+import { setSpanDurationAsMeasurement, setSpanDurationAsMeasurementOnSpan } from './utils';
 
 let nativeComponentMissingLogged = false;
 
@@ -206,14 +207,26 @@ function onDrawNextFrame(event: { nativeEvent: RNSentryOnDrawNextFrameEvent }): 
   }
 }
 
-function updateInitialDisplaySpan(frameTimestampSeconds: number): void {
-  const span = startTimeToInitialDisplaySpan();
+/**
+ *
+ */
+export function updateInitialDisplaySpan(
+  frameTimestampSeconds: number,
+  {
+    activeSpan = getActiveSpan(),
+    span = startTimeToInitialDisplaySpan(),
+  }: {
+    activeSpan?: Span;
+    /**
+     * Time to initial display span to update.
+     */
+    span?: Span;
+  } = {}): void {
   if (!span) {
     logger.warn(`[TimeToDisplay] No span found or created, possibly performance is disabled.`);
     return;
   }
 
-  const activeSpan = getActiveSpan();
   if (!activeSpan) {
     logger.warn(`[TimeToDisplay] No active span found to attach ui.load.initial_display to.`);
     return;
@@ -239,7 +252,7 @@ function updateInitialDisplaySpan(frameTimestampSeconds: number): void {
     updateFullDisplaySpan(frameTimestampSeconds, span);
   }
 
-  setSpanDurationAsMeasurement('time_to_initial_display', span);
+  setSpanDurationAsMeasurementOnSpan('time_to_initial_display', span, activeSpan);
 }
 
 function updateFullDisplaySpan(frameTimestampSeconds: number, passedInitialDisplaySpan?: Span): void {
@@ -283,4 +296,56 @@ function updateFullDisplaySpan(frameTimestampSeconds: number, passedInitialDispl
   logger.debug(`[TimeToDisplay] ${spanJSON.description} (${spanJSON.span_id}) span updated with end timestamp.`);
 
   setSpanDurationAsMeasurement('time_to_full_display', span);
+}
+
+/**
+ * Creates a new TimeToFullDisplay component which triggers the full display recording every time the component is focused.
+ */
+export function createTimeToFullDisplay({
+  useFocusEffect,
+}: {
+  /**
+   * `@react-navigation/native` useFocusEffect hook.
+   */
+  useFocusEffect: (callback: () => void) => void
+}): React.ComponentType<TimeToDisplayProps> {
+  return createTimeToDisplay({ useFocusEffect, Component: TimeToFullDisplay });
+}
+
+/**
+ * Creates a new TimeToInitialDisplay component which triggers the initial display recording every time the component is focused.
+ */
+export function createTimeToInitialDisplay({
+  useFocusEffect,
+}: {
+  useFocusEffect: (callback: () => void) => void
+}): React.ComponentType<TimeToDisplayProps> {
+  return createTimeToDisplay({ useFocusEffect, Component: TimeToInitialDisplay });
+}
+
+function createTimeToDisplay({
+  useFocusEffect,
+  Component,
+}: {
+  /**
+   * `@react-navigation/native` useFocusEffect hook.
+   */
+  useFocusEffect: (callback: () => void) => void;
+  Component: typeof TimeToFullDisplay | typeof TimeToInitialDisplay;
+}): React.ComponentType<TimeToDisplayProps> {
+  const TimeToDisplayWrapper = (props: TimeToDisplayProps): React.ReactElement => {
+    const [focused, setFocused] = useState(false);
+
+    useFocusEffect(() => {
+        setFocused(true);
+        return () => {
+          setFocused(false);
+        };
+    });
+
+    return <Component {...props} record={focused && props.record} />;
+  };
+
+  TimeToDisplayWrapper.displayName = `TimeToDisplayWrapper`;
+  return TimeToDisplayWrapper;
 }