+++ packages/core/src/js/utils/rnlibraries.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
 
-import { Platform, TurboModuleRegistry } from 'react-native';
+import { AppRegistry, Platform, TurboModuleRegistry } from 'react-native';
 
 import type * as ReactNative from '../vendor/react-native';
 import type { ReactNativeLibrariesInterface } from './rnlibrariesinterface';
@@ -34,6 +34,7 @@ export const ReactNativeLibraries: Required<ReactNativeLibrariesInterface> = {
     version: Platform.constants?.reactNativeVersion,
   },
   TurboModuleRegistry,
+  AppRegistry,
   ReactNative: {
     requireNativeComponent: <T>(viewName: string): ReactNative.HostComponent<T> => {
       const { requireNativeComponent } = require('react-native');